临时开放wuhaolun的供应商付款审批权限。

+++ database/migrations/2020_11_07_164906_create_epcseceningcranes_table.php
@@ -17,7 +17,7 @@ class CreateEpcseceningcranesTable extends Migration
 
             $table->integer('epcsecening_id')->unsigned();
             $table->string('crane_type');
-            $table->integer('number')->default(0);
+            $table->decimal('number')->default(0);
             $table->decimal('unitprice')->default(0);
 //            $table->decimal('totalprice',8,2);
